refactor: 의미에 맞게 메서드명 변경
